+++ agents/imp_builder/evals/files/helpers.py
@@ -1,0 +1,3 @@
+"""Out of scope bait. Leave `_tmp` alone."""
+
+_tmp = 0